Move icon name getter to cross-platform header
